Fix some sagemaker bugs: - not creating session with profile - not taking up the cluster if no address found in check_server - Don't fail if role or profile are not found to create session when on the cluster
Remove "resources" from function, which is obsolete.

Don't let env save if it has name Env.DEFAULT_NAME. This clearly needs more work (e.g. not setting the default name directly, but rather implicitly) but for now is ok.

Fix alt_options so if an option is specified which is not in the config, it counts as a miss.