Fix test_from_config to correctly use existing env. (#190)
Passing an env as a part of `.to` overrides the entire existing env initialized as part of the function definition. This is, for now, expected behavior.

Moreover, when initializing from_config, we need to explicitly set the working_dir else it'll be set to None, and the current function will not be pushed up. 

**Testing**: `pytest -v --level local -k "TestFunction"`