Fix the env syncing to replicas in the mapper
Previously, after the env was sent to the cluster (all nodes by default) in the main module,
we were syncing the env's packages to the cluster again for each replica's .to. This was actually
a bug having to do with the cluster not being recognized when recreating the package objects within
the env. To fix it, we 1) don't recreat the cluster when we duplicate the env for replicas 2) don't
strip the env from the module before replicating it (which cases installs in the replicas to fail)
and 3) update how folder recreates a system from config to be consistent with the rest of the repo
(not sure this makes a difference but worth doing anyway).