Overhaul put_resource logic. (#356)
Introduce a generic format and precedence for calls into `EnvServlet`.

All functions will take in `data` and `serialization`. They can put whatever
they want in the `data` field and unpack them as tuples before and after, as
done in this PR. This will support pickle and json serialization.

The other parameters to `EnvServlet` functions will be explicit.

The `EnvServlet` should also only call `local` ObjStore functions, and the
ObjStore should deal with all the outside-of-env logic and call the right Env.