Simplify `rh.here` Ray cluster detection logic. (#390)
The logic that we previously had to detect if we are on a cluster was extra
tricky because we would initializing the `rh.here` variable upon `import
runhouse as rh`, so there would be circular situations where we need to avoid
deadlocks in Actor creation.

However, now this is almost always being called in a process where one should
easily be able to connect to Ray and get the `cluster_servlet` if it exists. We
don't need any fancy `list_actors` nonsense or an extra `ray.init`.

This is also part of a more general theme of trying to trim down the excess
`ray.init`s that we have in the code to simplify the `ray start` sequence.