Fix `put_resource` to not store resource in itself if it is an Env. (#410)
If we put an env, we ended up storing the Env resource within the Env servlet
for that env, which doesn't really make sense. Changed put_resource to be
similar to put, where you can choose where the resource is stored, and if it
happens to be an env then we create that servlet, however don't necessarily
store it.

You can still simulate that behavior by adding the logic back on the client
side (in cluster.put_resource) that sets env_name to the env if it is one.