Only set rns_address upon save
Until now we've been setting the _rns_folder of a resource implicitly to
rns_client.current_folder, whether the resource has been saved or not,
so the address is really meaningless. This is a relic from a time when
we tried to save resources constantly as they were created and changed.
Now, it would be much easier to have a reasonable assumption of whether
a resource has been saved and therefore whether its rns_address is
meaningful if passed to another environment. For example, when saving
a resource config as a sub-config of another resource, we need to know
whether we can just save a string because the resource has been saved
or if we need to return the full sub-resource config. This bug in particular
was breaking some things.

This change changes the behavior of `rns_address` to return None if there
is no _rns_folder or no _name, and saves the _rns_folder to the current_folder
only upon save. If a user gives a resource a full rns_address that's their own
business, we will not check if the address is actually pointing to something
when we return the property.