RNS client error should match the new modified log string.
We now catch in `handle_exception_response` a `PermissionError` and check if
the text is `No Runhouse Token Provided.` in order to raise the appropriate
HTTPException error. This was not being caught because the log string wasn't
updated, so therefore the error was getting raised on the server and causing
a 500.