Test Telemetry Accross other Test Containers
Previously we created a dedicated container for testing whether telemetry is enabled in LOCAL.
We should be testing this and our ability to access spans over HTTP accross many of our test
container types (e.g. access over ssh with and w/o den auth, over HTTP and TLS). We'll
also now be testing that we can't access spans for any cluster which hasn't enabled it,
including the non-LOCAL cluster types.