Change requests.Session scope to be global
Turns out requests.Session only pools HTTP connections within a session:
https://github.com/psf/requests/issues/3445#issuecomment-263820399

We were creating new connections to the same cluster for no reason. Now we use
a global pool for the process. It would be nice to pool across processes, but
not sure that's straightforward, as requests is barely even threadsafe.