Change all `serialize_data` and `deserialize_data` spots to take lists instead of tuples. (#485)
Tuples aren't json serializable. Therefore when we are passing arguments around that could be serialized as json (since we support None, json, pickle), we should have them packed as Lists. When we unpack these, we cast them to tuples and use normal tuple unwrapping methods; this allows us to still check the arguments are in the form we expect.