Change some logging to go to `sys.stdout`. (#519)
When calling a method, we use the `StreamTee` in `provenance.py` to log to a file _and_ stdout at the same time, which is how we stream logs from `call`.

However, when we call `env.install`, we call things via subprocess, and these logs aren't propagated. Unfortunately there's no neat way to pipe this (I scoured the entirety of StackOverflow) so we just need to set up a pipe, stream the logs, and print things, and then they'll go to `sys.stdout` and be picked up by the remote log generator. 

While I was here, cleaned up a bit of the conda installation logic.