Keep stub fns `remote_await` and `remote_anext` sync. (#688)
Either these functions `await` a coroutine and are coroutines themselves, or are sync functions that return a coroutine, that is awaited in `acall_local` anyway. Semantically, it doesn't matter. However, the function signature remaining synchronous is useful for the async PR up in my stack. Changing this first to detect any breakages that are just on this part.