Don't serialize tracebacks in OutputType.EXCEPTION Responses (#721)
Previously, we serialized both the exception and the traceback together, so if the deserialization failed (e.g. if there was a library incompatibility between the client and server which blocked unpickling) we couldn't see the traceback either, even though it's a string. It's important that we try to desrialize the exception so users have a chance to catch it properly locally, but if it's impossible to deserialize, we may as well just raise a runtime error and still print the traceback. If the response from the cluster is still in the old format, we'll still try to unpickle it for compatibility, but we can eventually deprecate that.