Change ObjStore methods to call by env servlet name instead of actor. (#733)
We're seeing many cases where the EnvServlet dies, but it isn't cleaned up from global state. Then, if the user re-runs their code (which may once again call put resource, since it calls `.to`), it'll think the Servlet is already on the machine (because of our global `env_servlets` dict) and not re-create it. This makes the server fully borked after an EnvServlet dies. This is basic fault tolerance to clean up objects if this happens.