Rewrite Module tests to be more thorough and separate. (#727)
Right now we have `test_shared_readonly` which is the main test that was checking inter cluster calling flows and calling after sharing. This test was being skipped in release testing because `TEST_TOKEN` and `TEST_USERNAME` were not set. 

Now, it fails because the underlying `AsyncGenerator` that we use to mock `async for` calls is not shared with the user. I've renamed this. I've renamed this to `test_share_module_with_generator_read_only_inter_cluster`

Creating a more simple tests that checks `share` and `revoke` with calling a simple Module.

- `export TEST_TOKEN=...`
- `export TEST_USERNAME=...`
- `pytest -v --level local -k "test_share_module_read_only"`
- `pytest -v --level release -k "test_share_module_read_only_inter_cluster"`
- `pytest -v --level local "test_construct_shared_module"`
- **FAILING CURRENTLY**: `pytest -v --level release "test_share_module_with_generator_read_only_inter_cluster"`