Don't replace `base_env` with `system.default_env`. (#751)
There are several instances where, if the env happened to be `base_env`, we replace it with the cluster default env. The problem is we don't always replace this client side and it creates some inconsistencies. Wanted to discuss this logic a bit more

I've marked these with `TODO: (default env)`

cc: @carolineechen when you're back