Change AutostopServlet into AutostopHelper, and test properly (#897)
AutostopServlet is no longer in its own process because it's easier to just run
python commands in SkyPilot's venv through subprocess than start an actor inside
that interpreter. Here we're changing the AutostopServlet to AutostopHelper (a regular class,
not an Actor), fleshing out the available autostop commands it can make,  and testing it thoroughly.

Previously keep_warm or the cluster.autostop_mins setter wouldn't work if called on the cluster,
now they do.