conda_env in local Docker cluster (#1053)
Add a new dockerfile with miniconda installed to be able to test conda default_env in a local docker cluster. Unfortunately changing the dockerfile to use the miniconda base image did not work for the other local clusters which don't use conda default env, not for lack of trying. Ideally we can consolidate these somehow soon.